Split out an API module
This allows dependent projects that are only interested in the API
(such as submariner-addon) to only import the API.

Fixes: #1542
