+++ apis/go.mod
@@ -1,0 +1,10 @@
+module github.com/submariner-io/submariner-operator/apis
+
+go 1.15
+
+require (
+	github.com/submariner-io/submariner/pkg/apis v0.0.0-20210816153739-c8e6654e3930
+	k8s.io/api v0.19.10
+	k8s.io/apimachinery v0.19.10
+	sigs.k8s.io/controller-runtime v0.7.0
+)
