+++ apis/submariner/v1alpha1/versions.go
@@ -15,7 +15,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package versions
+package v1alpha1
 
 var (
 	DefaultRepo                      = "quay.io/submariner"
