+++ pkg/images/images.go
@@ -21,8 +21,8 @@ import (
 	"fmt"
 	"strings"
 
+	apis "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/pkg/names"
-	"github.com/submariner-io/submariner-operator/pkg/versions"
 
 	v1 "k8s.io/api/core/v1"
 )
@@ -72,7 +72,7 @@ func ParseOperatorImage(operatorImage string) (string, string) {
 	imageName := strings.Replace(operatorImage, repository, "", 1)
 	i := strings.LastIndex(imageName, ":")
 	if i == -1 {
-		version = versions.DefaultSubmarinerOperatorVersion
+		version = apis.DefaultSubmarinerOperatorVersion
 	} else {
 		version = imageName[i+1:]
 	}
